requirements.txt
python-dotenv==1.0.0
Requests==2.31.0
ta==0.10.3
yfinance==0.2.18
